Removal of SafeMath and Errors and other
Removal of SafeMath (not needed anymore) and Errors.sol (not doing that
kind of optimizations now).

Added protection in the maximum number of eTokens allowed, to avoid an
addEToken attach.

Also fixed an error in test_riskmodule

Finally, rearranged Policy struct to use less words of storage and being
more gas efficient.