New parameter for RiskModules: MOC
MOC = Margin Of Conversativism is a parameter that affects purePremium
calculation.
Before, purePremium = premium * lossProb
Now, purePremium = premium * lossProb * moc

moc defaults to 1. It can be adjusted to >1 if we want to reserve more
premium for pure premium pool if lossProb might be underestimated. Also
it can be adjusted to <1 if we want to reserve less premium.