New Pricing - interest rate fixed / RM variable
New pricing strategy. Now each RiskModule has a fixed interest rate that
has to pay for the capital locked (SCR).

Also renamed `ensuro_share` to `ensuro_fee` and now is calculated as a
percentage of pure premium.

Finally, removed premium_share (the share for RM), because now that is
variable, is whatever exceeds the pure_premium + ensuro_fee + capital
cost.