Insolvency Hook
This component of the PolicyPool will have the strategy to implement
when on an insolvency situation.

It has only one method `outOfCash` and should add money to the pool some
way.

Along with this change I'm adding the `receiveGrant` method to the
PolicyPool. This method receives _free_ cash from the sender and adds
that cash to the premium pool.