Insolvency Hook that add liquidity to EToken
This required several changes and a refactoring in EToken and
PolicyPool.

This kind of insolvency hook can work providing liquidity to one of the
ETokens. Can be the same EToken used for the policy being resolved or
any other.

Now poolLoan can receive a greater amount than possible and returns the
amount of loan taken.

Also, in the scale_factor I added a protection to keep it far from zero
to avoid zero division errors and rounding errors if too low.

When the loan can't be taken from tokens associated to the policy, is
taken from the other tokens.