Remove Shared Coverage feature
We were supporting an option of "shared coverage", where the risk module
covers a given percentage of the payout and takes the same percentage of
the premium. That payout was taken upfront from the risk partner,
requiring full collateralization on their side.

We are not sure this will be used in practice, so I removed it to
simplify the code and reduce gas cost.