Change in SCR computation
Previously the scr was:

`scr = (payout - premium) * scr_percentage`

This wasn't correct because even defining the `scr_percentage` as
`(VAR995-VAR50)/VAR100`, it wasn't the right value.

With that definition for the scr_percentage the right formula should be

`scr = payout * scr_percentage` (assuming premium == VAR50, that's not
true in general).

Instead of that, now the `scr_percentage` should be defined as
`VAR995/VAR100` (the proportion of the current colateralization level
with respect to full colateralization).

And the `scr` is now defined as:

`scr = payout * scr_percentage - premium`