Deterministic Policy Id
Before this change, the policyId number it's known only after the
newPolicy transaction is processed, because it was a counter in the
PolicyNFT contract.

This was a problem for partners integrating with us that need to avoid
duplicated policies and need to match their internal id with the
Ensuro's Policy Id.

With this change, the user needs to send a `internalId` parameter and
the resulting policyId is created merging this internalId with the Risk
Module address:

`policyId = uint256(riskModule.address) << 96 + internalId`

This way we can know in advance the policyId and at the same time we
avoid colission of the numbers between two different modules.