Removes horizonal SCR split
Until now, the SCR of a policy was split between several eTokens. This
was adding a lot of complexity that's not really needed, since most of
the time just one eToken accepts a policy.

Now, each policy has a single solvency eToken. We still keep the
expiration limit at the eToken level, but now it's not so relevant.

Work in progress, change done only in the Python code, missing the
Solidity Code.