QSP-17 Reverts max approval - Approves when needed
Previous code was doing a max approval for both eTokens when this
approval is only needed when the premium account has to repay a loan.

This was exposing PremiumsAccount to any possible compromise on eToken
contracts.

Instead now we delay the approval until we have to repay a loan. In that
case the approval is done for all the pending debt, with the expectation
to avoid some approvals in the future.