Whitelisting per operation
Previously we couldn't blacklist the withdrawals. Now we added another
check in the eToken calling `acceptsWithdrawal`.

In the implementation of the LPManualWhitelist we added more
granularity. Now for each LP we can specify different whitelisting
status for each of the four operations: deposit, withdraw, sending
transfers, receiving transfers.

The LPManualWhitelist also receives the default behaviours for each of
the operations if the LP doesn't have anything defined.