Migration contract for 2.0 whitelist
Migration contract that will be used to upgrade the current 2.0
whitelist. This just implements the `acceptsWithdrawal` endpoint using
acceptsDeposit.

No tests included here, it will be tested in the deployment repository.

This contract will be used only for a few days until a brand new
whitelist contract is deployed and the whitelisted users are migrated.