PolicyPool migration to custom errors
Most of the errors in the PolicyPool were migrated to custom errors.
This reduces the size of the contract (that was very close to the
limit). Also, is a step towards full migration to custom errors that we
will do in v3.

Policy errors were left as strings to not break bridge23 or other
internal projects that use the error messages, and the Python tests.