Pluggable IAssetManager strategy => ERC4626 yield vault
Removal of the IAssetManager pluggable strategy that was plugged into
the reserves (EToken and PremiumsAccount) and called with delegate call
and instead we now have just an ERC4626 yield vault.

Also, we remove the rebalancing policy that was based on thresholds and
now the rebalancing logic (when to rebalance between liquid and
invested) is a responsability of the caller.

Removes references to the asset manager in the Python code and tests. I
will test this with Javascript tests.

Also I removed the NEGLIGIBLE_AMOUNT hack. The issue (if/when happens)
can be solved by gifting a few pennies to the reserve.