Flexible Policy Replacements
Change in Policy Replacements. Now we accept changes in the exposure
(payout), solvency amounts and also a shorter expiration.

What we don't accept is reduction in any of the components of the
premium.

If the interest rate changes, it produces a discrete adjustment in the
eToken, to reflect the difference in the interest accrued with the old
policy and the new policy.

Just the Solidity code was changed. There's one breaking test left
intentionally broken.

In the next commit I will fix that test and include others.