Option cooldown for EToken withdrawals
Posibility to set a `cooler()` contract that requires scheduling the
withdrawals before executing them.

The scheduled withdrawals can be executed on the deadline, and can
return an amount lower or equal to the amount requested.

If there were losses, the LP suffers the losses.

If there are profits, the profits are distributed to the remaining LPs.