ENSR-4 Fix infinit approval to YieldVault
As recommended by the auditors, we changed the infinite approval given
to the yieldVault every time it is plugged (setYieldVault), and now the
approval is made on the `depositIntoYieldVault` method and should be spend by the `deposit` call on the vault.

This way, a malicious vault can only drain the funds it received, but
not all the funds of the reserves.